refactor(config): move config related files to their own folder
