+++ src/Application.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "CommandManager.h"
-#include "ConfigurationValueTree.h"
+#include "config/ConfigurationValueTree.h"
 #include "gui/appearance/LookAndFeel.h"
 #include "gui/windows/MainWindow.h"
 #include <JuceHeader.h>
