+++ src/config/CMakeLists.txt
@@ -1,0 +1,14 @@
+###############################################################################
+#                   :
+#   File            :   CMakeLists.txt
+#                   :
+#   Author(s)       :   Tim Brewis (@t-bre)
+#                   :
+#   Description     :   CMakeLists for config  model
+#                   :
+###############################################################################
+
+target_sources(${TARGET_NAME}
+    PRIVATE 
+        ConfigurationValueTree.cpp
+)
